pyproject.toml
[project]
name = "bugzilla-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for interacting with Bugzilla bug tracking systems"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "Apache-2.0"}
keywords = ["mcp", "bugzilla", "bug-tracking", "model-context-protocol", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Bug Tracking",
]

dependencies = [
    "fastmcp==2.12.5",
    "python-dotenv",
    "httpx",
]
